requirements.txt
Babel==1.3
argparse==1.2.1
iso8601==0.1.10
msgpack-python==0.4.6
netaddr==0.7.14
netifaces==0.10.4
oslo.config==1.11.0
oslo.i18n==1.6.0
oslo.serialization==1.5.0
oslo.utils==1.5.0
pbr==0.11.0
prettytable==0.7.2
python-cinderclient==1.2.1
python-keystoneclient==1.4.0
pytz==2015.4
requests==2.7.0
simplejson==3.6.5
six==1.9.0
stevedore==1.4.0
wsgiref==0.1.2
